more cleaning up the codebase and normalizing things (#383)
* more cleaning up the codebase and normalizing things

* tons more of the cleanup/normaliztions

* same...

* more cleanup and a small bug fix for a crash when the window height is literally 0

* more of the same...

* minor stuff

* the names of settings aren't really meant to be translated, just use QLatin1String

* more cleanup and simplification

* using QStringLiteral in a bunch more places

* more ...

* more normalization of style

* moving the macro compilation one step closer to being fully memory mananged

* removing a nolonger needed comment

* fixing some compiler errors

* some use of QStringLiteral in a few more places

* more mior optimizations

* lots of better QStringLiteral usage

* formating/style/etc

* more normalization, cleanup and docs stuff

* more doc stuff

* squash some qt6 warnings when cranked up

* clean up some more warnings when turned up

* simplified some code

* squash some more windows warnings

* more warning stuff

* more of the same...

* font-merging breaks the font being truely 'fixed width'

+++ Regex/Reader.h
@@ -267,7 +267,7 @@ public:
 
 			auto index = std::distance(str.begin(), match[0].first);
 			auto size  = std::distance(match[0].first, match[0].second);
-			std::basic_string_view<Ch> m(str.data() + index, size);
+			std::basic_string_view<Ch> m(str.data() + index, static_cast<size_t>(size));
 
 			index_ += m.size();
 			return m;
